// resources/js/api/generated/models/operationLogStatsData.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface OperationLogStatsData {
  failed_operations: number;
  incomplete_operations: number;
  recent_operations: number;
  successful_operations: number;
  total_operations: number;
}